#365 validate lookup attributes on concept create/edit (#374)
