+++ django-nonrel/ocl/concepts/validation_messages.py
@@ -1,9 +1,3 @@
-BASIC_DESCRIPTION_LOCALE = 'Description locale must be one of the valid choices'
-BASIC_NAME_LOCALE = 'Name locale must be one of the valid choices'
-BASIC_DESCRIPTION_TYPE = 'Description type must be one of the valid choices'
-BASIC_NAME_TYPE = 'Name type must be one of the valid choices'
-BASIC_DATATYPE = 'Data type must be valid attribute'
-BASIC_CONCEPT_CLASS = 'Concept class must be valid attribute'
 BASIC_DESCRIPTION_CANNOT_BE_EMPTY = 'Concept description cannot be empty'
 
 OPENMRS_ONE_FULLY_SPECIFIED_NAME_PER_LOCALE = 'A concept may not have more than one fully specified name in any locale'
@@ -14,5 +8,11 @@ OPENMRS_MUST_HAVE_EXACTLY_ONE_PREFERRED_NAME = 'A concept may not have more than
 OPENMRS_SHORT_NAME_CANNOT_BE_PREFERRED = 'A short name cannot be marked as locale preferred'
 OPENMRS_AT_LEAST_ONE_FULLY_SPECIFIED_NAME = 'A concept must have at least one fully specified name (across all locales)'
 OPENMRS_PREFERRED_NAME_UNIQUE_PER_SOURCE_LOCALE = 'Concept preferred name must be unique for same source and locale'
+OPENMRS_DESCRIPTION_LOCALE = 'Description locale must be one of the valid choices'
+OPENMRS_NAME_LOCALE = 'Name locale must be one of the valid choices'
+OPENMRS_DESCRIPTION_TYPE = 'Description type must be one of the valid choices'
+OPENMRS_NAME_TYPE = 'Name type must be one of the valid choices'
+OPENMRS_DATATYPE = 'Data type must be valid attribute'
+OPENMRS_CONCEPT_CLASS = 'Concept class must be valid attribute'
